# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Supabase Auth API",
    "version": "1.0.0",
    "description": "Secure API with Supabase Auth, JWT verification, and protected routes"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Supabase JWT access token obtained from /auth/login"
      }
    },
    "schemas": {
      "AuthRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "developer@flyrank.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "SuperSecretPass123!"
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": {
            "type": "string",
            "example": "d290f1ee6c1746c6..."
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" },
          "token_type": { "type": "string", "example": "bearer" },
          "expires_in": { "type": "integer", "example": 3600 },
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "email": { "type": "string" },
              "created_at": { "type": "string" }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/public/info": {
      "get": {
        "summary": "Public open information",
        "description": "Accessible by anyone without authentication",
        "responses": {
          "200": { "description": "Public greeting message" }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Register a new user account",
        "description": "Creates an account with Supabase Auth Identity Provider",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in and obtain JWT access token (Rate limited)",
        "description": "Authenticates user credentials against Supabase and returns JWT. Protected with 429 brute-force limit.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful with JWT access token" },
          "400": { "description": "Missing credentials" },
          "401": { "description": "Invalid login credentials" },
          "429": { "description": "Too many failed attempts (Locked out for 5 minutes)" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Refresh expired access token",
        "description": "Rotates refresh token and returns fresh access token without re-prompting for credentials",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Token rotated successfully" },
          "400": { "description": "Missing refresh token" },
          "401": { "description": "Invalid or expired refresh token" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out user and end session",
        "description": "Protected endpoint that invalidates the user's active session",
        "security": [
          { "BearerAuth": [] }
        ],
        "responses": {
          "204": { "description": "Logged out successfully (No Content)" },
          "401": { "description": "Missing, malformed, or invalid token" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Read authenticated user profile",
        "description": "Guarded by Supabase JWT verification middleware",
        "security": [
          { "BearerAuth": [] }
        ],
        "responses": {
          "200": { "description": "Authenticated user profile details" },
          "401": { "description": "Missing, malformed, or invalid token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Read protected dashboard data",
        "description": "Demonstrates reuse of the same auth middleware guard across multiple endpoints",
        "security": [
          { "BearerAuth": [] }
        ],
        "responses": {
          "200": { "description": "Protected user metrics and dashboard view" },
          "401": { "description": "Missing, malformed, or invalid token" }
        }
      }
    },
    "/protected/admin": {
      "get": {
        "summary": "Admin-only restricted portal (Role-based Authorization)",
        "description": "Requires both authentication (401) and administrative role authorization (403)",
        "security": [
          { "BearerAuth": [] }
        ],
        "responses": {
          "200": { "description": "Admin metrics accessed successfully" },
          "401": { "description": "Unauthenticated user" },
          "403": { "description": "Forbidden: User lacks administrative authorization" }
        }
      }
    }
  }
}